# docs/swagger.yaml
definitions:
...
  requests.GenerateAnswer:
    properties:
      condition:
        maxLength: 2000
        type: string
    required:
    - condition
    type: object
  requests.GenerateByInterests:
    properties:
      condition:
        maxLength: 2000
        type: string
      interests:
        items:
          type: string
        maxItems: 20
        minItems: 0
        type: array
    required:
    - condition
    - interests
    type: object
  requests.GenerateByNoInterests:
    properties:
      condition:
        maxLength: 2000
        type: string
    required:
    - condition
    type: object
...
  responses.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  responses.GeneratedAnswerResponse:
    properties:
      generated_text:
        type: string
      status:
        type: string
    type: object
  responses.GeneratedTaskResponse:
    properties:
      generated_text:
        type: string
      status:
        type: string
    type: object
...
  responses.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/generate/answer:
    post:
      consumes:
      - application/json
      description: Generates an answer based on the provided condition and saves it
        in history.
      parameters:
      - description: Data for answer generation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.GenerateAnswer'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated answer
          schema:
            $ref: '#/definitions/responses.GeneratedAnswerResponse'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Error saving the generated answer
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Generate Answer by Condition
      tags:
      - Answer Generation
  /api/generate/interests:
    post:
      consumes:
      - application/json
      description: Generates a task based on the provided list of interests and saves
        it in the history.
      parameters:
      - description: Data for task generation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.GenerateByInterests'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated task
          schema:
            $ref: '#/definitions/responses.GeneratedTaskResponse'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Error saving the generated task
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Generate Task by Interests
      tags:
      - Task Generation
  /api/generate/nointerests:
    post:
      consumes:
      - application/json
      description: Generates a task based solely on the provided condition and saves
        it in history.
      parameters:
      - description: Data for task generation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.GenerateByNoInterests'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated task
          schema:
            $ref: '#/definitions/responses.GeneratedTaskResponse'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Error saving the generated task
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Generate Task Without Interests
      tags:
      - Task Generation
...
swagger: "2.0"